Fix some merge errors
